// src/ApiModule/model/program.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstructorPublic } from './instructorPublic';
import { Category } from './category';
import { Placement } from './placement';


export interface Program { 
    name: string;
    category: Category;
    placement: Placement;
    instructor: InstructorPublic;
    paid: boolean;
    place_limit?: number;
    registration_opens?: number;
    available_registration?: boolean;
    id: number;
}