// src/ApiModule/model/createProgram.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstructorPublic } from './instructorPublic';
import { Category } from './category';
import { Placement } from './placement';


export interface CreateProgram {
    name: string;
    category: Category | string;
    placement: Placement | string;
    instructor: InstructorPublic | number;
    paid: boolean;
    place_limit?: number;
    registration_opens?: number;
    available_registration?: boolean;
}